# docs/v1/swagger.yaml
basePath: /api/v1
definitions:
  dto.CreateAnswerRequest:
    properties:
      content:
        minLength: 10
        type: string
      user_id:
        type: string
    required:
    - content
    - user_id
    type: object
...
  dto.UpdateAnswerRequest:
    properties:
      content:
        minLength: 10
        type: string
    type: object
...
  response.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorInfo'
      message:
        type: string
      meta:
        $ref: '#/definitions/response.MetaInfo'
      success:
        type: boolean
      timestamp:
        type: string
    type: object
  response.ErrorInfo:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
  response.MetaInfo:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: A RESTful API built with Go and Fiber
  title: Pitstop API
  version: "1.0"
paths:
...
  /questions/{question_id}/answers:
    get:
      consumes:
      - application/json
      description: Retrieve all answers for a specific question
      parameters:
      - description: Question ID
        in: path
        name: question_id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Answers per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Get answers for a question
      tags:
      - answers
    post:
      consumes:
      - application/json
      description: Create an answer for a specific question
      parameters:
      - description: Question ID
        in: path
        name: question_id
        required: true
        type: string
      - description: Answer details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAnswerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create an answer
      tags:
      - answers
  /questions/{question_id}/answers/{answer_id}:
    delete:
      consumes:
      - application/json
      description: Delete an answer (only by author)
      parameters:
      - description: Question ID
        in: path
        name: question_id
        required: true
        type: string
      - description: Answer ID
        in: path
        name: answer_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete an answer
      tags:
      - answers
    get:
      consumes:
      - application/json
      description: Retrieve a specific answer
      parameters:
      - description: Question ID
        in: path
        name: question_id
        required: true
        type: string
      - description: Answer ID
        in: path
        name: answer_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Get an answer by ID
      tags:
      - answers
    put:
      consumes:
      - application/json
      description: Update an answer (only by author)
      parameters:
      - description: Question ID
        in: path
        name: question_id
        required: true
        type: string
      - description: Answer ID
        in: path
        name: answer_id
        required: true
        type: string
      - description: Updated answer details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAnswerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update an answer
      tags:
      - answers
  /questions/{question_id}/answers/{answer_id}/accept:
    post:
      consumes:
      - application/json
      description: Mark an answer as accepted (only by question author)
      parameters:
      - description: Question ID
        in: path
        name: question_id
        required: true
        type: string
      - description: Answer ID
        in: path
        name: answer_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Accept an answer
      tags:
      - answers
  /questions/{question_id}/answers/{answer_id}/unaccept:
    post:
      consumes:
      - application/json
      description: Mark an answer as not accepted (only by question author)
      parameters:
      - description: Question ID
        in: path
        name: question_id
        required: true
        type: string
      - description: Answer ID
        in: path
        name: answer_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Unaccept an answer
      tags:
      - answers
...
swagger: "2.0"